requirements.txt
# Dependencias principales de Flask
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2

# Extensiones de Flask
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Bootstrap==3.3.7.1
Flask-Migrate==4.0.5
email-validator==2.1.0.post1

# Bases de datos y ORM
SQLAlchemy==2.0.21
pymysql==1.1.0
cryptography==41.0.4  # Para conexiones seguras a MySQL
alembic==1.12.0

# Formularios y validación
WTForms==3.0.1

# Manejo de variables de entorno
python-dotenv==1.0.0

# Utilidades
requests==2.31.0
Pillow==10.0.1  # Para procesamiento de imágenes si se necesitan

# Seguridad
bcrypt==4.0.1  # Para hash de contraseñas

# Extras para desarrollo
pytest==7.4.2
pytest-flask==1.2.0
flake8==6.1.0
